a2a-mesh/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2a-mesh"
version = "0.1.0"
description = "Local agent star network with web UI, built on A2A protocol"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"

dependencies = [
    # A2A SDK
    "a2a-sdk>=0.2.0",
    # LLM Client
    "openai>=1.0.0",
    "litellm>=1.0.0",
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "starlette>=0.36.0",
    # Network discovery
    "zeroconf>=0.131.0",
    # HTTP client
    "httpx>=0.27.0",
    # Config
    "pyyaml>=6.0",
    "pydantic>=2.0",
    # File watching
    "watchfiles>=0.21.0",
    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[project.scripts]
mesh = "mesh.cli:main"
